requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Requests==2.32.3
talisman==0.1.0
urllib3==1.26.18
